Fix tests
Tests in 5.1, sending CONTINUATION in closed or half-closed (remote)
state, are controversial.  For instance, 5.1 says receiving frames
other than WINDOW_UPDATE, PRIORITY or RST_STREAM in half-closed
(remote) must be treated as STREAM_CLOSED, but 6.10 also says that

"""
A CONTINUATION frame MUST be preceded by a HEADERS, PUSH_PROMISE or
CONTINUATION frame without the END_HEADERS flag set.  A recipient
that observes violation of this rule MUST respond with a connection
error (Section 5.4.1) of type PROTOCOL_ERROR.
"""

END_HEADERS flag set implies END_STREAM flag set.  So receiving
CONTINUATION after END_HEADERS set and END_STREAM set (half-closed
(remote)) must be treated as PROTOCOL_ERROR.

They seem contradict each other.  So I suggest to add PROTOCOL_ERROR
as acceptable error code for these tests.

The fix in 5.10 (half-closed (remote)) is avoid the implementation to
emit COMPRESSION_ERROR, because we set END_HEADERS in HEADERS frame,
which means we have to supply complete header block to it.

For test in 6.10, we must not close stream with END_STREAM flag set in
DATA.  If we do that, it is completely valid for the implementation to
emit STREAM_CLOSED error code.  To fix this, remove END_STREAM flag.
This ensures that implementation will emit PROTOCOL_ERROR for
unexpected CONTINUATION.